// src/types/canvas-cursor-types.ts
export type LineOptions = {
  spring: number;
};

export type Position = {
  x: number;
  y: number;
};

export type OscillatorOptions = {
  phase?: number;
  offset?: number;
  frequency?: number;
  amplitude?: number;
};

export type NodeType = {
  x: number;
  y: number;
  vx: number;
  vy: number;
};